Add files via upload
Used to give prompts to the Llama API it gives two prompts. The first prompt is used to enhance achievements and results without any structure. The second prompt summarizes the information of first prompt and returns an enhanced summary of the text. The file is used to structure the text elements, boxes and color them using reportlab. It uses Groq and Langchain for communication with AI, parsing ,prompts,etc.